update animation state
