+++ packages/space-opera/src/components/animation_controls/reducer.ts
@@ -16,18 +16,29 @@
  */
 
 import {Action} from '../../types.js';
+import {AnimationInfo} from './types.js';
 
-/** Set playAnimation or not */
+const SET_ANIMATION_NAMES = 'SET_ANIMATION_NAMES';
+export function dispatchSetAnimationNames(animationNames: string[]) {
+  return {type: SET_ANIMATION_NAMES, payload: animationNames};
+}
+
 const PLAY_ANIMATION = 'PLAY_ANIMATION';
 export function dispatchPlayAnimation(playAnimation: boolean) {
   return {type: PLAY_ANIMATION, payload: playAnimation};
 }
 
-export function playAnimationReducer(
-    state: boolean|undefined = true, action: Action): boolean|undefined {
+export function animationInfoReducer(
+    state: AnimationInfo = {
+      animationNames: [],
+      playAnimation: true
+    },
+    action: Action): AnimationInfo {
   switch (action.type) {
     case PLAY_ANIMATION:
-      return !!action.payload;
+      return {...state, playAnimation: !!action.payload};
+    case SET_ANIMATION_NAMES:
+      return {...state, animationNames: action.payload};
     default:
       return state;
   }
